xml5ever

Module serialize

Serializer for XML5.

Structs

NamespaceMapStack 🔒

SerializeOpts

Struct for setting serializer options.

XmlSerializer

Struct used for serializing nodes into a text that other XML parses can read.

Enums

TraversalScope

Used as a parameter to serialize, telling it if we want to skip the parent.

Traits

Serialize

Types that can be serialized (according to the xml-like scheme in Serializer) implement this trait.

Serializer

Types that are capable of serializing implement this trait

Functions

serialize

Method for serializing generic node to a given writer.

write_qual_name 🔒

write_to_buf_escaped 🔒

Writes given text into the Serializer, escaping it, depending on where the text is written inside the tag or attribute value.

Type Aliases

AttrRef

A type alias for an attribute name and value (e.g. the class="test" in <div class="test"> is represented as (<QualName of type class>, "test").