Struct xml5ever::serialize::XmlSerializer

pub struct XmlSerializer<Wr> { /* fields omitted */ }

Struct used for serializing nodes into a text that other XML parses can read.

Serializer contains a set of functions (start_elem, end_elem...) that make parsing nodes easier.

Methods

impl<Wr: Write> XmlSerializer<Wr>

pub fn new(writer: Wr) -> Self

Creates a new Serializier from a writer and given serialization options.