Struct xml5ever::tokenizer::QName

pub struct QName {
    pub prefix: Atom,
    pub local: Atom,
    pub namespace_url: Atom,
}

Fully qualified name. Used to depict names of tags and attributes.

Used to differentiate between similar XML fragments. For example ```ignore // HTML

Apples 	Bananas

// Furniture XML
African Coffee Table 80 120
``` Without XML namespaces we can't use those two fragments in occur XML at same time. however if we declare a namespace we could instead say:

   // Furniture XML
   <furn:table>
     <furn:name>African Coffee Table</furn:name>
     <furn:width>80</furn:width>
     <furn:length>120</furn:length>
   </furn:table>

and bind it to a different name.

For this reason we parse names that contain a colon in the following way

   < furn:table>
       |    |
       |    +- local name
       |
     prefix (when resolved gives namespace_url)

Fields

prefix

Prefix of fully qualified name, used for namespace lookup.

local

Local name of a value.

namespace_url

Resolved namespace of QName.

Methods

impl QName

fn new(prefix: Atom, local: Atom) -> QName

Constructs a new QName from prefix and local part. Namespace is set to empty atom.

fn new_empty(local: Atom) -> QName

Constructs a new QName with only local part. Namespace is set to empty atom.