Module xml5ever::tree_builder

XML5 tree builder - converts tokens into a tree like structure

Modules

types

Structs

NamespaceMapStack

XmlTreeBuilder
The XML tree builder.

XmlTreeBuilderOpts
Tree builder options, with an impl for Default.

Enums

NextParserState
Whether to interrupt further parsing of the current input until the next explicit resumption of the tokenizer, or continue without any interruption.

NodeOrText
Something which can be inserted into the DOM.

Statics

XMLNS_URI

XML_URI

Traits

Tracer
Trace hooks for a garbage-collected DOM.

TreeSink
Methods a parser can use to create the DOM. The DOM provider implements this trait.

Functions

any_not_whitespace

current_node

Type Definitions

InsResult