Trait xml5ever::tree_builder::actions::XmlTreeBuilderActions

pub trait XmlTreeBuilderActions<Handle> {
    fn current_node(&self) -> Handle;
    fn insert_appropriately(&mut self, child: NodeOrText<Handle>);
    fn insert_tag(&mut self, tag: Tag) -> XmlProcessResult;
    fn append_tag(&mut self, tag: Tag) -> XmlProcessResult;
    fn append_tag_to_doc(&mut self, tag: Tag) -> Handle;
    fn add_to_open_elems(&mut self, el: Handle) -> XmlProcessResult;
    fn append_comment_to_doc(&mut self, comment: StrTendril) -> XmlProcessResult;
    fn append_comment_to_tag(&mut self, text: StrTendril) -> XmlProcessResult;
    fn append_doctype_to_doc(&mut self, doctype: Doctype) -> XmlProcessResult;
    fn append_pi_to_doc(&mut self, pi: Pi) -> XmlProcessResult;
    fn append_pi_to_tag(&mut self, pi: Pi) -> XmlProcessResult;
    fn append_text(&mut self, chars: StrTendril) -> XmlProcessResult;
    fn tag_in_open_elems(&self, tag: &Tag) -> bool;
    fn pop_until<TagSet>(&mut self, pred: TagSet) where TagSet: Fn(QName) -> bool;
    fn current_node_in<TagSet>(&self, set: TagSet) -> bool where TagSet: Fn(QName) -> bool;
    fn close_tag(&mut self, tag: Tag) -> XmlProcessResult;
    fn no_open_elems(&self) -> bool;
    fn pop(&mut self) -> Handle;
    fn stop_parsing(&mut self) -> XmlProcessResult;
    fn complete_script(&mut self);
}

Trait that encapsulates common XML tree actions.

Required Methods

fn current_node(&self) -> Handle

Returns current node of in the XmlTreeBuilder.

fn insert_appropriately(&mut self, child: NodeOrText<Handle>)

Inserts node or text to its appropriate place in the tree.

fn insert_tag(&mut self, tag: Tag) -> XmlProcessResult

Inserts tag into the tree and adds it to list of open elements.

fn append_tag(&mut self, tag: Tag) -> XmlProcessResult

Appends current tag to the root of the document.

fn append_tag_to_doc(&mut self, tag: Tag) -> Handle

Appends tag to the root of the document.

fn add_to_open_elems(&mut self, el: Handle) -> XmlProcessResult

Adds element to list of open elements (this should only apply to Tag).

fn append_comment_to_doc(&mut self, comment: StrTendril) -> XmlProcessResult

Appends comment to root of the document.

fn append_comment_to_tag(&mut self, text: StrTendril) -> XmlProcessResult

Appends comment to the current tag.

fn append_doctype_to_doc(&mut self, doctype: Doctype) -> XmlProcessResult

Appends Doctype to root of the document.

fn append_pi_to_doc(&mut self, pi: Pi) -> XmlProcessResult

Appends Processing Instruction to the root of the document

fn append_pi_to_tag(&mut self, pi: Pi) -> XmlProcessResult

Appends Processing Instruction to the current tag.

fn append_text(&mut self, chars: StrTendril) -> XmlProcessResult

Appends text to appropriate element.

fn tag_in_open_elems(&self, tag: &Tag) -> bool

Checks if given tag is the list of open elements.

fn pop_until<TagSet>(&mut self, pred: TagSet) where TagSet: Fn(QName) -> bool

Pops elements from list of open elements, until predicate pred returns true

fn current_node_in<TagSet>(&self, set: TagSet) -> bool where TagSet: Fn(QName) -> bool

Checks if current node is in given TagSet

fn close_tag(&mut self, tag: Tag) -> XmlProcessResult

Close given tag.

fn no_open_elems(&self) -> bool

Returns whether or not there are any elements in list of open elements.

fn pop(&mut self) -> Handle

Removes last element from list of open elements and returns its value.

fn stop_parsing(&mut self) -> XmlProcessResult

Stops parsing of XML file.

fn complete_script(&mut self)

Indicated a script element is complete and can be prepared