xml5ever::tree_builder

Struct XmlTreeBuilder

pub struct XmlTreeBuilder<Handle, Sink> {
    _opts: XmlTreeBuilderOpts,
    pub sink: Sink,
    doc_handle: Handle,
    open_elems: RefCell<Vec<Handle>>,
    curr_elem: RefCell<Option<Handle>>,
    namespace_stack: RefCell<NamespaceMapStack>,
    current_namespace: RefCell<NamespaceMap>,
    phase: Cell<XmlPhase>,
}

The XML tree builder.

Fields
_opts: XmlTreeBuilderOpts

Configuration options for XmlTreeBuilder

sink: Sink

Consumer of tree modifications.

doc_handle: Handle

The document node, which is created by the sink.

open_elems: RefCell<Vec<Handle>>

Stack of open elements, most recently added at end.

curr_elem: RefCell<Option<Handle>>

Current element pointer.

namespace_stack: RefCell<NamespaceMapStack>

Stack of namespace identifiers and namespaces.

current_namespace: RefCell<NamespaceMap>

Current namespace identifier

phase: Cell<XmlPhase>

Current tree builder phase.

Implementations

impl<Handle, Sink> XmlTreeBuilder<Handle, Sink>
where Handle: Clone, Sink: TreeSink<Handle = Handle>,

pub fn new(sink: Sink, opts: XmlTreeBuilderOpts) -> XmlTreeBuilder<Handle, Sink>

Create a new tree builder which sends tree modifications to a particular TreeSink.

The tree builder is also a TokenSink.

pub fn trace_handles(&self, tracer: &dyn Tracer<Handle = Handle>)

Call the Tracer’s trace_handle method on every Handle in the tree builder’s internal state. This is intended to support garbage-collected DOMs.

fn dump_state(&self, label: String)

fn debug_step(&self, mode: XmlPhase, token: &Token)

fn declare_ns(&self, attr: &mut Attribute)

fn find_uri( &self, prefix: &Option<Prefix>, ) -> Result<Option<Namespace>, Cow<'static, str>>

fn bind_qname(&self, name: &mut QualName)

fn bind_attr_qname( &self, present_attrs: &mut HashSet<(Namespace, LocalName)>, name: &mut QualName, ) -> bool

fn check_duplicate_attr( present_attrs: &mut HashSet<(Namespace, LocalName)>, name: &QualName, ) -> bool

fn process_namespaces(&self, tag: &mut Tag)

fn process_to_completion( &self, token: Token, ) -> ProcessResult<<Self as TokenSink>::Handle>

impl<Handle, Sink> XmlTreeBuilder<Handle, Sink>
where Handle: Clone, Sink: TreeSink<Handle = Handle>,

fn current_node(&self) -> Ref<'_, Handle>

fn insert_appropriately(&self, child: NodeOrText<Handle>)

fn insert_tag(&self, tag: Tag) -> XmlProcessResult<Handle>

fn append_tag(&self, tag: Tag) -> XmlProcessResult<Handle>

fn append_tag_to_doc(&self, tag: Tag) -> Handle

fn add_to_open_elems(&self, el: Handle) -> XmlProcessResult<Handle>

fn append_comment_to_doc(&self, text: StrTendril) -> XmlProcessResult<Handle>

fn append_comment_to_tag(&self, text: StrTendril) -> XmlProcessResult<Handle>

fn append_doctype_to_doc(&self, doctype: Doctype) -> XmlProcessResult<Handle>

fn append_pi_to_doc(&self, pi: Pi) -> XmlProcessResult<Handle>

fn append_pi_to_tag(&self, pi: Pi) -> XmlProcessResult<Handle>

fn append_text(&self, chars: StrTendril) -> XmlProcessResult<Handle>

fn tag_in_open_elems(&self, tag: &Tag) -> bool

fn pop_until<P>(&self, pred: P)
where P: Fn(ExpandedName<'_>) -> bool,

fn current_node_in<TagSet>(&self, set: TagSet) -> bool
where TagSet: Fn(ExpandedName<'_>) -> bool,

fn close_tag(&self, tag: Tag) -> XmlProcessResult<Handle>

fn no_open_elems(&self) -> bool

fn pop(&self) -> Handle

fn stop_parsing(&self) -> XmlProcessResult<Handle>

impl<Handle, Sink> XmlTreeBuilder<Handle, Sink>
where Handle: Clone, Sink: TreeSink<Handle = Handle>,

fn step( &self, mode: XmlPhase, token: Token, ) -> XmlProcessResult<<Self as TokenSink>::Handle>