Struct xml5ever::tree_builder::XmlTreeBuilderOpts

pub struct XmlTreeBuilderOpts {
    pub exact_errors: bool,
    pub profile: bool,
}

Tree builder options, with an impl for Default.

Fields
exact_errors: bool

Report all parse errors described in the spec, at some performance penalty? Default: false

profile: bool

Keep a record of how long we spent in each state? Printed when end() is called. Default: false