Crate xml5ever

This crate provides a push based XML parser library that adheres to XML5 specification. In other words this library trades well-formedness for error recovery.

The idea behind this, was to minimize number of errors from tools that generate XML (e.g. &#83 won't just return &#83 as text, but will parse it into S ). You can check out full specification here.

What this library provides is a solid XML parser that can:

Parse somewhat erroneous XML input

Provide support for Numeric character references.

Provide partial XML namespace support.

Provide full set of SVG/MathML entities

What isn't in scope for this library:

Document Type Definition parsing - this is pretty hard to do right and nowadays, its used

Reexports

pub use markup5ever::*;

Modules

driver
Driver

serialize
Serializer for XML5.

tokenizer
XML5 tokenizer - converts input into tokens

tree_builder
XML5 tree builder - converts tokens into a tree like structure

Macros

expanded_name

local_name

namespace_prefix

namespace_url

ns

small_char_set
Create a SmallCharSet, with each space-separated number stored in the set.