Crate xml5ever

This crate provides a push based XML parser library that adheres to XML5 specification. In other words this library trades well-formedness for error recovery.

The idea behind this, was to minimize number of errors from tools that generate XML (e.g. &#83 won't just return &#83 as text, but will parse it into S ). You can check out full specification here.

What this library provides is a solid XML parser that can:

Parse somewhat erroneous XML input

Provide support for Numeric character references.

Provide partial XML namespace support.

Provide full set of SVG/MathML entities

What isn't in scope for this library:

Document Type Definition parsing - this is pretty hard to do right and nowadays, its used

Modules

driver
Entrance to XML5 ever ParseResult

encoding
Re-export the rust-encoding crate.

rcdom
A simple reference-counted that serves as a default tree structure A simple reference-counted DOM.

serialize
Serializer for XML5.

tendril
Re-export the tendril crate so that users don’t need to depend on it.

tokenizer
XML5 tokenizer - converts input into tokens

tree_builder
XML5 tree builder - converts tokens into a tree like structure

Structs

QualName
A name with a namespace.

Type Definitions

LocalName

Namespace

Prefix