Module xml5ever::tokenizer

XML5 tokenizer - converts input into tokens

Re-exports

pub use self::interface::CharacterTokens;

pub use self::interface::EOFToken;

pub use self::interface::NullCharacterToken;

pub use self::interface::CommentToken;

pub use self::interface::DoctypeToken;

pub use self::interface::PIToken;

pub use self::interface::TagToken;

pub use self::interface::EmptyTag;

pub use self::interface::EndTag;

pub use self::interface::ShortTag;

pub use self::interface::StartTag;

pub use self::interface::ParseError;

Modules

char_ref

interface

qname

states

Tokenizer states.

Structs

Doctype

A DOCTYPE token. Doctype token in XML5 is rather limited for reasons, such as: security and simplicity. XML5 only supports declaring DTD with name, public identifier and system identifier

Pi

A ProcessingInstruction token.

Tag

XML 5 Tag Token

XmlTokenizer

The Xml tokenizer.

XmlTokenizerOpts

Copy of Tokenizer options, with an impl for Default.

Enums

TagKind

Tag kind denotes which kind of tag did we encounter.

Token

Describes tokens encountered during parsing of input.

Traits

TokenSink

Types which can receive tokens from the tokenizer.

Functions

option_push

process_qname

Type Definitions

LocalName

Namespace

Prefix