Struct xml5ever::tokenizer::XmlTokenizer

pub struct XmlTokenizer<Sink> {
    opts: XmlTokenizerOpts,
    sink: Sink,
    state: XmlState,
    input_buffers: BufferQueue,
    at_eof: bool,
    char_ref_tokenizer: Option<Box<CharRefTokenizer>>,
    current_char: char,
    reconsume: bool,
    ignore_lf: bool,
    discard_bom: bool,
    current_tag_kind: TagKind,
    current_tag_name: StrTendril,
    current_tag_attrs: Vec<Attribute>,
    current_attr_name: StrTendril,
    current_attr_value: StrTendril,
    current_doctype: Doctype,
    current_comment: StrTendril,
    current_pi_target: StrTendril,
    current_pi_data: StrTendril,
    state_profile: BTreeMap<XmlState, u64>,
    time_in_sink: u64,
}

The Xml tokenizer.

Fields

opts

Options controlling the behavior of the tokenizer.

sink

Destination for tokens we emit.

state

The abstract machine state as described in the spec.

input_buffers

Input ready to be tokenized.

at_eof

Are we at the end of the file, once buffers have been processed completely? This affects whether we will wait for lookahead or not.

char_ref_tokenizer

Tokenizer for character references, if we're tokenizing one at the moment.

current_char

Current input character. Just consumed, may reconsume.

reconsume

Should we reconsume the current input character?

ignore_lf

Did we just consume \r, translating it to \n? In that case we need to ignore the next character if it's \n.

discard_bom

Discard a U+FEFF BYTE ORDER MARK if we see one? Only done at the beginning of the stream.

current_tag_kind

Current tag kind.

current_tag_name

Current tag name.

current_tag_attrs

Current tag attributes.

current_attr_name

Current attribute name.

current_attr_value

Current attribute value.

current_doctype

current_comment

Current comment.

current_pi_target

Current processing instruction target.

current_pi_data

Current processing instruction value.

state_profile

Record of how many ns we spent in each state, if profiling is enabled.

time_in_sink

Record of how many ns we spent in the token sink.

Methods

impl<Sink: TokenSink> XmlTokenizer<Sink>

fn new(sink: Sink, opts: XmlTokenizerOpts) -> XmlTokenizer<Sink>

Create a new tokenizer which feeds tokens to a particular TokenSink.

fn unwrap(self) -> Sink

Returns destination of token events.

fn sink<'a>(&'a self) -> &'a Sink

Immutably borrows destination of token events.

fn sink_mut<'a>(&'a mut self) -> &'a mut Sink

Mutably borrows destinantion of token events.

fn feed(&mut self, input: StrTendril)

Feed an input string into the tokenizer.

fn process_token(&mut self, token: Token)

fn get_preprocessed_char(&mut self, c: char) -> Option<char>

fn bad_eof_error(&mut self)

fn pop_except_from(&mut self, set: SmallCharSet) -> Option<SetResult>

fn eat(&mut self, pat: &str) -> Option<bool>

fn run(&mut self)

Run the state machine for as long as we can.

fn get_char(&mut self) -> Option<char>

fn bad_char_error(&mut self)

fn discard_tag(&mut self)

fn create_tag(&mut self, kind: TagKind, c: char)

fn create_pi(&mut self, c: char)

fn emit_char(&mut self, c: char)

fn emit_short_tag(&mut self)

fn emit_empty_tag(&mut self)

fn set_empty_tag(&mut self)

fn emit_start_tag(&mut self)

fn emit_current_tag(&mut self)

fn emit_chars(&mut self, b: StrTendril)

fn emit_pi(&mut self)

fn consume_char_ref(&mut self, addnl_allowed: Option<char>)

fn emit_eof(&mut self)

fn emit_error(&mut self, error: Cow<'static, str>)

fn emit_current_comment(&mut self)

fn emit_current_doctype(&mut self)

fn doctype_id<'a>(&'a mut self, kind: DoctypeKind) -> &'a mut Option<StrTendril>

fn clear_doctype_id(&mut self, kind: DoctypeKind)

fn peek(&mut self) -> Option<char>

fn discard_char(&mut self)

fn unconsume(&mut self, buf: StrTendril)

impl<Sink: TokenSink> XmlTokenizer<Sink>

fn step(&mut self) -> bool

fn end(&mut self)

Indicate that we have reached the end of the input.

fn dump_profile(&self)

fn eof_step(&mut self) -> bool

fn process_char_ref(&mut self, char_ref: CharRef)

fn step_char_ref_tokenizer(&mut self) -> bool

fn finish_attribute(&mut self)

fn create_attribute(&mut self, c: char)