Struct xml5ever::tokenizer::XmlTokenizer

pub struct XmlTokenizer<Sink> {
    pub sink: Sink,
    // some fields omitted
}

The Xml tokenizer.

Fields
sink: Sink

Destination for tokens we emit.

Methods

impl<Sink: TokenSink> XmlTokenizer<Sink>

pub fn new(sink: Sink, opts: XmlTokenizerOpts) -> XmlTokenizer<Sink>

Create a new tokenizer which feeds tokens to a particular TokenSink.

pub fn feed(&mut self, input: StrTendril)

Feed an input string into the tokenizer.

pub fn run(&mut self)

Run the state machine for as long as we can.

impl<Sink: TokenSink> XmlTokenizer<Sink>

pub fn end(&mut self)

Indicate that we have reached the end of the input.