Struct xml5ever::tokenizer::XmlTokenizer

pub struct XmlTokenizer<Sink> {
    // some fields omitted
}

The Xml tokenizer.

Methods

impl<Sink: XTokenSink> XmlTokenizer<Sink>

fn new(sink: Sink, opts: XmlTokenizerOpts) -> XmlTokenizer<Sink>

Create a new tokenizer which feeds tokens to a particular TokenSink.

fn unwrap(self) -> Sink

fn sink<'a>(&'a self) -> &'a Sink

fn sink_mut<'a>(&'a mut self) -> &'a mut Sink

fn feed(&mut self, input: StrTendril)

Feed an input string into the tokenizer.

impl<Sink: XTokenSink> XmlTokenizer<Sink>

fn end(&mut self)

Indicate that we have reached the end of the input.