Trait xml5ever::tokenizer::TokenSink

pub trait TokenSink {
    fn process_token(&mut self, token: Token);

    fn end(&mut self) { ... }

    fn query_state_change(&mut self) -> Option<XmlState> { ... }
}

Types which can receive tokens from the tokenizer.

Required methods

fn process_token(&mut self, token: Token)

Process a token.

Loading content...
Provided methods

fn end(&mut self)

Signal to the sink that parsing has ended.

fn query_state_change(&mut self) -> Option<XmlState>

The tokenizer will call this after emitting any start tag. This allows the tree builder to change the tokenizer's state. By default no state changes occur.

Loading content...