Trait xml5ever::tokenizer::TokenSink

pub trait TokenSink {
    // Required method
    fn process_token(&self, token: Token);

    // Provided methods
    fn end(&self) { ... }
    fn query_state_change(&self) -> Option<XmlState> { ... }
}

Types which can receive tokens from the tokenizer.

Required Methods

fn process_token(&self, token: Token)

Process a token.

Provided Methods

fn end(&self)

Signal to the sink that parsing has ended.

fn query_state_change(&self) -> Option<XmlState>

The tokenizer will call this after emitting any start tag. This allows the tree builder to change the tokenizer’s state. By default no state changes occur.