Struct xml5ever::tree_builder::XmlTreeBuilder

pub struct XmlTreeBuilder<Handle, Sink> {
    sink: Sink,
    doc_handle: Handle,
    next_tokenizer_state: Option<XmlState>,
    open_elems: Vec<Handle>,
    curr_elem: Option<Handle>,
    namespace_stack: NamespaceStack,
    current_namespace: Namespace,
    present_attrs: HashSet<(Atom, Atom)>,
    phase: XmlPhase,
}

The XML tree builder.

Fields

sink

Consumer of tree modifications.

doc_handle

The document node, which is created by the sink.

next_tokenizer_state

Next state change for the tokenizer, if any.

open_elems

Stack of open elements, most recently added at end.

curr_elem

Current element pointer.

namespace_stack

Stack of namespace identifiers and namespaces.

current_namespace

Current namespace identifier

present_attrs

List of already present namespace local name attribute pairs.

phase

Current tree builder phase.

Methods

impl<Handle, Sink> XmlTreeBuilder<Handle, Sink> where Handle: Clone, Sink: TreeSink<Handle=Handle>

fn new(sink: Sink) -> XmlTreeBuilder<Handle, Sink>

Create a new tree builder which sends tree modifications to a particular TreeSink.

The tree builder is also a TokenSink.

fn unwrap(self) -> Sink

Returns consumer of tree modifications.

fn sink<'a>(&'a self) -> &'a Sink

Immutably borrows consumer of tree modifications.

fn sink_mut<'a>(&'a mut self) -> &'a mut Sink

Mutably borrows consumer of tree modifications.

fn trace_handles(&self, tracer: &Tracer<Handle=Handle>)

Call the Tracer's trace_handle method on every Handle in the tree builder's internal state. This is intended to support garbage-collected DOMs.

fn dump_state(&self, label: String)

fn debug_step(&self, mode: XmlPhase, token: &Token)

fn declare_ns(&mut self, attr: &mut Attribute)

fn find_uri(&self, prefix: &Atom) -> Result<Option<Atom>, Cow<'static, str>>

fn bind_qname(&mut self, name: &mut QName)

fn bind_attr_qname(&mut self, name: &mut QName) -> bool

fn check_duplicate_attr(&mut self, name: &QName) -> bool

fn process_namespaces(&mut self, tag: &mut Tag)

fn process_to_completion(&mut self, token: Token)