Struct xml5ever::tree_builder::XmlTreeBuilder

pub struct XmlTreeBuilder<Handle, Sink> {
    pub sink: Sink,
    // some fields omitted
}

The XML tree builder.

Fields
sink: Sink

Consumer of tree modifications.

Methods

impl<Handle, Sink> XmlTreeBuilder<Handle, Sink> where
Handle: Clone,
Sink: TreeSink<Handle = Handle>,

pub fn new(sink: Sink, opts: XmlTreeBuilderOpts) -> XmlTreeBuilder<Handle, Sink>

Create a new tree builder which sends tree modifications to a particular TreeSink.

The tree builder is also a TokenSink.

pub fn trace_handles(&self, tracer: &Tracer<Handle = Handle>)

Call the Tracer's trace_handle method on every Handle in the tree builder's internal state. This is intended to support garbage-collected DOMs.