Struct xml5ever::tree_builder::XmlTreeBuilder

pub struct XmlTreeBuilder<Handle, Sink> {
    _opts: XmlTreeBuilderOpts,
    pub sink: Sink,
    doc_handle: Handle,
    next_tokenizer_state: Cell<Option<XmlState>>,
    open_elems: RefCell<Vec<Handle>>,
    curr_elem: RefCell<Option<Handle>>,
    namespace_stack: RefCell<NamespaceMapStack>,
    current_namespace: RefCell<NamespaceMap>,
    phase: Cell<XmlPhase>,
}

The XML tree builder.

Fields
_opts: XmlTreeBuilderOpts

Configuration options for XmlTreeBuilder

sink: Sink

Consumer of tree modifications.

doc_handle: Handle

The document node, which is created by the sink.

next_tokenizer_state: Cell<Option<XmlState>>

Next state change for the tokenizer, if any.

open_elems: RefCell<Vec<Handle>>

Stack of open elements, most recently added at end.

curr_elem: RefCell<Option<Handle>>

Current element pointer.

namespace_stack: RefCell<NamespaceMapStack>

Stack of namespace identifiers and namespaces.

current_namespace: RefCell<NamespaceMap>

Current namespace identifier

phase: Cell<XmlPhase>

Current tree builder phase.

Implementations

impl<Handle, Sink> XmlTreeBuilder<Handle, Sink>
where Handle: Clone, Sink: TreeSink<Handle = Handle>,

pub fn new(sink: Sink, opts: XmlTreeBuilderOpts) -> XmlTreeBuilder<Handle, Sink>

Create a new tree builder which sends tree modifications to a particular TreeSink.

The tree builder is also a TokenSink.

pub fn trace_handles(&self, tracer: &dyn Tracer<Handle = Handle>)

Call the Tracer’s trace_handle method on every Handle in the tree builder’s internal state. This is intended to support garbage-collected DOMs.

fn dump_state(&self, label: String)

fn debug_step(&self, mode: XmlPhase, token: &Token)

fn declare_ns(&self, attr: &mut Attribute)

fn find_uri( &self, prefix: &Option<Prefix>, ) -> Result<Option<Namespace>, Cow<'static, str>>

fn bind_qname(&self, name: &mut QualName)

fn bind_attr_qname( &self, present_attrs: &mut HashSet<(Namespace, LocalName)>, name: &mut QualName, ) -> bool

fn check_duplicate_attr( present_attrs: &mut HashSet<(Namespace, LocalName)>, name: &QualName, ) -> bool

fn process_namespaces(&self, tag: &mut Tag)

fn process_to_completion(&self, token: Token)