Struct xml5ever::tree_builder::XmlTreeBuilder

pub struct XmlTreeBuilder<Handle, Sink> {
    _opts: XmlTreeBuilderOpts,
    pub sink: Sink,
    doc_handle: Handle,
    next_tokenizer_state: Option<XmlState>,
    open_elems: Vec<Handle>,
    curr_elem: Option<Handle>,
    namespace_stack: NamespaceMapStack,
    current_namespace: NamespaceMap,
    present_attrs: HashSet<(Namespace, LocalName)>,
    phase: XmlPhase,
}

The XML tree builder.

Fields
_opts: XmlTreeBuilderOpts

Configuration options for XmlTreeBuilder

sink: Sink

Consumer of tree modifications.

doc_handle: Handle

The document node, which is created by the sink.

next_tokenizer_state: Option<XmlState>

Next state change for the tokenizer, if any.

open_elems: Vec<Handle>

Stack of open elements, most recently added at end.

curr_elem: Option<Handle>

Current element pointer.

namespace_stack: NamespaceMapStack

Stack of namespace identifiers and namespaces.

current_namespace: NamespaceMap

Current namespace identifier

present_attrs: HashSet<(Namespace, LocalName)>

List of already present namespace local name attribute pairs.

phase: XmlPhase

Current tree builder phase.

Implementations

impl<Handle, Sink> XmlTreeBuilder<Handle, Sink> where
Handle: Clone,
Sink: TreeSink<Handle = Handle>,

pub fn new(sink: Sink, opts: XmlTreeBuilderOpts) -> XmlTreeBuilder<Handle, Sink>

Create a new tree builder which sends tree modifications to a particular TreeSink.

The tree builder is also a TokenSink.

pub fn trace_handles(&self, tracer: &dyn Tracer<Handle = Handle>)

Call the Tracer’s trace_handle method on every Handle in the tree builder’s internal state. This is intended to support garbage-collected DOMs.

fn dump_state(&self, label: String)

fn debug_step(&self, mode: XmlPhase, token: &Token)

fn declare_ns(&mut self, attr: &mut Attribute)

fn find_uri(
&self,
prefix: &Option<Prefix>
) -> Result<Option<Namespace>, Cow<'static, str>>

fn bind_qname(&mut self, name: &mut QualName)

fn bind_attr_qname(&mut self, name: &mut QualName) -> bool

fn check_duplicate_attr(&mut self, name: &QualName) -> bool

fn process_namespaces(&mut self, tag: &mut Tag)

fn process_to_completion(&mut self, token: Token)