Struct xml5ever::tree_builder::XmlTreeBuilder

pub struct XmlTreeBuilder<Handle, Sink> { /* fields omitted */ }

The XML tree builder.

Methods

impl<Handle, Sink> XmlTreeBuilder<Handle, Sink> where Handle: Clone,
Sink: TreeSink<Handle=Handle>

fn new(sink: Sink, opts: XmlTreeBuilderOpts) -> XmlTreeBuilder<Handle, Sink>

Create a new tree builder which sends tree modifications to a particular TreeSink.

The tree builder is also a TokenSink.

fn unwrap(self) -> Sink

Returns consumer of tree modifications.

fn sink<'a>(&'a self) -> &'a Sink

Immutably borrows consumer of tree modifications.

fn sink_mut<'a>(&'a mut self) -> &'a mut Sink

Mutably borrows consumer of tree modifications.

fn trace_handles(&self, tracer: &Tracer<Handle=Handle>)

Call the Tracer's trace_handle method on every Handle in the tree builder's internal state. This is intended to support garbage-collected DOMs.